Multilingual_AI_Assistance/requirements.txt
SpeechRecognition
pipwin
pyaudio

# google text to speech
gTTS
google-generativeai
python-dotenv
streamlit
